moved all test files to _test_ folder and rename the test files to same as normal comopnent name style
