+++ frontend/src/_test_/SignUpFields.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import SignUpFields from '../../Components/Users/signUp/SignUpFields';
+import SignUpFields from '../Components/Users/signUp/SignUpFields';
 
 configure({ adapter: new Adapter() });
