+++ frontend/src/_test_/SignUpForm.test.js
@@ -3,17 +3,17 @@ import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import reducerRoot from '../../reducers';
-import ForgotPassword from '../../Components/Users/forgot-password';
+import reducerRoot from '../reducers';
+import SignUpForm from '../Components/Users/signUp';
 
 configure({ adapter: new Adapter() });
 
 const store = createStore(reducerRoot);
 
-describe('ForgotPassword Component', () => {
+describe('SignUpForm Component', () => {
     it('Should render without crashing', () => {
         const wrapper = 
-        shallow(<Provider store={store}><ForgotPassword /></Provider>).length
+        shallow(<Provider store={store}><SignUpForm /></Provider>).length
         expect(wrapper).toEqual(1)
     })
 })
